Mapmatching is a mode. works on actions
